requirements.txt
discord-py-interactions
CopperUI
flask
discord
requests
mysql-connector-python